+++ backend/src/main/java/com/university/skillshare_backend/config/WebConfig.java
@@ -2,20 +2,13 @@ package com.university.skillshare_backend.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.web.multipart.MultipartResolver;
-import org.springframework.web.multipart.support.StandardServletMultipartResolver;
 
-import jakarta.annotation.PostConstruct;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 
 @Configuration
@@ -25,20 +18,6 @@ public class WebConfig implements WebMvcConfigurer {
 
     @Value("${file.upload-dir:./uploads}")
     private String uploadDir;
-
-    @PostConstruct
-    public void init() throws IOException {
-        Path uploadPath = Paths.get(uploadDir).toAbsolutePath().normalize();
-        Files.createDirectories(uploadPath);
-        Files.createDirectories(uploadPath.resolve("images"));
-        Files.createDirectories(uploadPath.resolve("videos"));
-        logger.info("Created upload directories at: {}", uploadPath);
-    }
-
-    @Bean
-    public MultipartResolver multipartResolver() {
-        return new StandardServletMultipartResolver();
-    }
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
